# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Categories Controller",
        "version": "1"
    },
    "paths": {
...
        "/admin/coupons/{id}": {
            "get": {
                "tags": [
                    "coupons"
                ],
                "operationId": "73a3b0477255de64025d58b328e800a4",
                "responses": {
                    "200": {
                        "description": "product fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminCoupon"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "coupons"
                ],
                "summary": "Remove the specified resource from storage.",
                "operationId": "27bba561cc9b8f63f2ecabf6db0d7be5",
                "responses": {
                    "204": {
                        "description": "The Coupon was successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "coupons"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "5c7da04233de229db03258455a733dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminUpdateCoupon"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Coupon created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminCouponIdPathParameter"
                }
            ]
        },
        "/admin/drivers/{id}": {
            "get": {
                "tags": [
                    "drivers"
                ],
                "operationId": "23991b377afeae29ae24b70cabd53aa8",
                "responses": {
                    "204": {
                        "description": "Fetched Driver Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminDriverData"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "drivers"
                ],
                "summary": "Remove the specified resource from storage.",
                "operationId": "988bff1b3534097d6ba1daa2f1f8f081",
                "responses": {
                    "204": {
                        "description": "The Driver was successfully deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "drivers"
                ],
                "summary": "Update the specified resource in storage.",
                "operationId": "0dc77c9fda7e61f464f6783353ece330",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminUpdateDriverData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Driver created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminDriverData"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "$ref": "#/components/parameters/adminDriverIdPathParameter"
                }
            ]
        },
...
        "/admin/coupons": {
            "get": {
                "tags": [
                    "coupons"
                ],
                "operationId": "428fbed59bb70eecc4149bd9e21957ae",
                "responses": {
                    "200": {
                        "description": "The Coupon was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adminCoupon"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "coupons"
                ],
                "operationId": "4063250b7b17361fffddb51cd1891c4d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminCreateCoupon"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Coupon created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/drivers": {
            "get": {
                "tags": [
                    "drivers"
                ],
                "summary": "Get All Drivers",
                "operationId": "3782754d223fdbf4799a436af6eface0",
                "responses": {
                    "200": {
                        "description": "The Driver was successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/adminDriverData"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "drivers"
                ],
                "summary": "Create a new Driver.",
                "operationId": "594841f04d3a8caac8ad17112c921189",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adminCreateDriverData"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Driver created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/adminDriverData"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "adminCoupon": {
                "properties": {
                    "id": {
                        "description": "CouponUserData>",
                        "type": "integer"
                    },
                    "name": {
                        "description": "CouponUserData>",
                        "type": "string"
                    },
                    "code": {
                        "description": "CouponUserData>",
                        "type": "string"
                    },
                    "percent": {
                        "description": "CouponUserData>",
                        "type": "string",
                        "nullable": true
                    },
                    "value": {
                        "description": "CouponUserData>",
                        "type": "string",
                        "nullable": true
                    },
                    "start_at": {
                        "description": "CouponUserData>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-08-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "end_at": {
                        "description": "CouponUserData>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2014-09-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "created_at": {
                        "description": "CouponUserData>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2017-02-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "users": {
                        "description": "CouponUserData>",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/adminCouponUser"
                        }
                    }
                },
                "type": "object"
            },
            "adminCouponUser": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "adminCreateCoupon": {
                "properties": {
                    "name": {
                        "description": "int>",
                        "type": "string"
                    },
                    "code": {
                        "description": "int>",
                        "type": "string"
                    },
                    "percent": {
                        "description": "int>",
                        "type": "string",
                        "nullable": true
                    },
                    "value": {
                        "description": "int>",
                        "type": "integer",
                        "nullable": true
                    },
                    "start_at": {
                        "description": "int>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-08-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "end_at": {
                        "description": "int>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-09-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "user_ids": {
                        "description": "int>",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "group_ids": {
                        "description": "int>",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
...
            "adminUpdateCoupon": {
                "properties": {
                    "name": {
                        "description": "int>",
                        "type": "string"
                    },
                    "code": {
                        "description": "int>",
                        "type": "string"
                    },
                    "percent": {
                        "description": "int>",
                        "type": "string",
                        "nullable": true
                    },
                    "value": {
                        "description": "int>",
                        "type": "integer",
                        "nullable": true
                    },
                    "start_at": {
                        "description": "int>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-08-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "end_at": {
                        "description": "int>",
                        "type": "string",
                        "format": "datetime",
                        "default": "2024-09-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    },
                    "user_ids": {
                        "description": "int>",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    },
                    "group_ids": {
                        "description": "int>",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "adminCreateDriverData": {
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "nullable": true
                    },
                    "number": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "adminDriverData": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "username": {
                        "type": "string",
                        "nullable": true
                    },
                    "number": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "default": "2017-02-02 18:31:45",
                        "pattern": "YYYY-MM-DD"
                    }
                },
                "type": "object"
            },
...
            "adminUpdateDriverData": {
                "properties": {
                    "name": {
                        "type": "string",
                        "nullable": true
                    },
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string",
                        "nullable": true
                    },
                    "number": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "adminCouponIdPathParameter": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
            "adminDriverIdPathParameter": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ]
}